+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Added `getStripeCustomerByEmail` on the Customers service layer
 - Added `getStripeCustomer` to the enupalStripe variable
+- Added `getPaymentMethod` to the Customers service layer
 
 ## 2.1.5 - 2019.12.05